chore: move token entry to core-service to allow reuse
