+++ jans-auth-server/server/src/main/java/io/jans/as/server/auth/AuthenticationFilter.java
@@ -25,13 +25,13 @@ import io.jans.as.server.model.common.AbstractToken;
 import io.jans.as.server.model.common.AuthorizationCodeGrant;
 import io.jans.as.server.model.common.AuthorizationGrant;
 import io.jans.as.server.model.common.AuthorizationGrantList;
-import io.jans.as.server.model.ldap.TokenEntity;
 import io.jans.as.server.model.token.ClientAssertion;
 import io.jans.as.server.model.token.HttpAuthTokenType;
 import io.jans.as.server.service.*;
 import io.jans.as.server.service.token.TokenService;
 import io.jans.as.server.util.TokenHashUtil;
 import io.jans.model.security.Identity;
+import io.jans.model.token.TokenEntity;
 import io.jans.service.CacheService;
 import io.jans.util.StringHelper;
 import jakarta.inject.Inject;
