+++ jans-auth-server/server/src/main/java/io/jans/as/server/model/common/IAuthorizationGrant.java
@@ -10,7 +10,7 @@ import io.jans.as.common.model.common.User;
 import io.jans.as.common.model.registration.Client;
 import io.jans.as.model.common.GrantType;
 import io.jans.as.server.model.authorize.JwtAuthorizationRequest;
-import io.jans.as.server.model.ldap.TokenEntity;
+import io.jans.model.token.TokenEntity;
 
 import java.util.Collection;
 import java.util.Date;
