+++ jans-auth-server/server/src/main/java/io/jans/as/server/service/CleanerTimer.java
@@ -17,12 +17,12 @@ import io.jans.as.model.uma.persistence.UmaResource;
 import io.jans.as.persistence.model.ClientAuthorization;
 import io.jans.as.persistence.model.Par;
 import io.jans.as.persistence.model.Scope;
-import io.jans.as.server.model.ldap.TokenEntity;
 import io.jans.as.server.uma.authorization.UmaPCT;
 import io.jans.as.server.uma.service.UmaPctService;
 import io.jans.as.server.uma.service.UmaResourceService;
 import io.jans.model.ApplicationType;
 import io.jans.model.metric.ldap.MetricEntry;
+import io.jans.model.token.TokenEntity;
 import io.jans.orm.PersistenceEntryManager;
 import io.jans.orm.search.filter.Filter;
 import io.jans.service.cache.CacheProvider;
