+++ jans-auth-server/server/src/test/java/io/jans/as/server/comp/CleanerTimerTest.java
@@ -14,10 +14,10 @@ import io.jans.as.server.BaseComponentTest;
 import io.jans.as.server.model.common.AccessToken;
 import io.jans.as.server.model.common.ClientCredentialsGrant;
 import io.jans.as.server.model.common.ExecutionContext;
-import io.jans.as.server.model.ldap.TokenEntity;
 import io.jans.as.server.model.token.HandleTokenFactory;
 import io.jans.as.server.uma.authorization.UmaPCT;
 import io.jans.as.server.uma.authorization.UmaRPT;
+import io.jans.model.token.TokenEntity;
 import io.jans.util.security.StringEncrypter;
 import jakarta.ws.rs.WebApplicationException;
 import org.testng.annotations.Test;
