+++ jans-core/service/src/main/java/io/jans/model/token/TokenEntity.java
@@ -4,7 +4,7 @@
  * Copyright (c) 2020, Janssen Project
  */
 
-package io.jans.as.server.model.ldap;
+package io.jans.model.token;
 
 import io.jans.as.model.common.GrantType;
 import io.jans.orm.annotation.AttributeName;
